refactor(refactor1): перенес компоненты в соотвецтвующие папки
