+++ src/components/empty-favorites/empty-favorites.tsx
