+++ src/components/favorite-card/favorite-card.tsx
