+++ src/pages/login/login.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../components/header';
+import Header from '../../components/header/header';
 
 const Login: React.FC = () => (
   <div className="page page--gray page--login">
